v1.2.0

+++ packages/artinchip/lvgl-ui/lvgl_v9/lvgl/src/osal/lv_rtthread.c
@@ -49,7 +49,7 @@ lv_result_t lv_thread_init(lv_thread_t * thread, lv_thread_prio_t prio, void (*c
                                       THREAD_TIMESLICE);
     rt_err_t ret = rt_thread_startup(thread->thread);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -61,7 +61,7 @@ lv_result_t lv_thread_delete(lv_thread_t * thread)
 {
     rt_err_t ret = rt_thread_delete(thread->thread);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -85,7 +85,7 @@ lv_result_t lv_mutex_lock(lv_mutex_t * mutex)
 {
     rt_err_t ret = rt_mutex_take(mutex->mutex, RT_WAITING_FOREVER);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -97,7 +97,7 @@ lv_result_t lv_mutex_lock_isr(lv_mutex_t * mutex)
 {
     rt_err_t ret = rt_mutex_take(mutex->mutex, RT_WAITING_FOREVER);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -109,7 +109,7 @@ lv_result_t lv_mutex_unlock(lv_mutex_t * mutex)
 {
     rt_err_t ret = rt_mutex_release(mutex->mutex);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -121,7 +121,7 @@ lv_result_t lv_mutex_delete(lv_mutex_t * mutex)
 {
     rt_err_t ret = rt_mutex_delete(mutex->mutex);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -145,7 +145,7 @@ lv_result_t lv_thread_sync_wait(lv_thread_sync_t * sync)
 {
     rt_err_t ret = rt_sem_take(sync->sem, RT_WAITING_FOREVER);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -157,7 +157,7 @@ lv_result_t lv_thread_sync_signal(lv_thread_sync_t * sync)
 {
     rt_err_t ret = rt_sem_release(sync->sem);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
@@ -169,7 +169,7 @@ lv_result_t lv_thread_sync_delete(lv_thread_sync_t * sync)
 {
     rt_err_t ret = rt_sem_delete(sync->sem);
     if(ret) {
-        LV_LOG_WARN("Error: %d", ret);
+        LV_LOG_WARN("Error: %d", (int)ret);
         return LV_RESULT_INVALID;
     }
     else {
